Calculate variance covariance parameters

Scale variance-covariance matrices as the relative Cholesky factors of each random effect term.

Usage

calc.theta(VarCov, sigma)

Arguments

VarCov

variance-covariance matrices. If there are multiple random effect groups, supply the variance-covariance matrix of each group as an element in a list.

sigma

standard deviation of random errors.

Value

theta

See also