Correlation Structure Classes

Standard classes of correlation structures (corStruct) available from the nlme package and re-exported by pwr4exp for convenience when specifying correlation structures in mkdesign.

All arguments are identical to the corresponding nlme functions. For more details on the original implementations, see nlme::corClasses.

Note: In the original nlme::corAR1, nlme::corARMA, and nlme::corSymm functions, the covariate t in the correlation formula ~ t or ~ t | g must be an integer class. In pwr4exp, the covariate can also be a factor class, which is then converted to an integer internally for sorting purposes. The class of the covariate variable in the model formula, if present, will not be converted. For example, a time covariate can be fitted as a factor in the model formula, whereas it is converted to an integer in the correlation formula temporarily for matrix sorting.

Details

Available standard classes:

corAR1

autoregressive process of order 1.

corARMA

autoregressive moving average process, with arbitrary orders for the autoregressive and moving average components.

corCAR1

continuous AR(1)

corCompSymm

compound symmetry structure corresponding to a constant correlation.

corExp

exponential spatial correlation.

corGaus

Gaussian spatial correlation.

corLin

linear spatial correlation.

corRatio

Rational quadratics spatial correlation.

corSpher

spherical spatial correlation.

corSymm

general correlation matrix, with no additional structure.

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.