Create a Customized Design

Create a Customized Design

Usage

designCustom(design.df, formula, beta, VarCov, sigma2, design.name, ...)

Arguments

design.df

a data frame containing values of all independent variables as columns

formula

an object of class formula: a symbolic description of the model that would be used to test effects in post-experimental data analysis. For details on model specification, see lm for fixed effects and lmer for random effects.

beta

a vector of the expectations of model coefficients.The coefficients for categorical variables are the coefficients of dummy variables created using contr.treatment contrast coding.

VarCov

variance-covariance components of random effects. If there are multiple random effects of one grouping factor, provide the variance-covariance matrix. If there are multiple grouping factors, supply the variance-covariance matrix of each grouping factor in a list.

sigma2

error variance

design.name

optional, name of the design

...

other arguments passed to the anova function in lmerTest. The type of ANOVA table, with Type III as the default, and the method for computing the denominator degrees of freedom, with Satterthwaite's method as the default, can be changed. For more details, see anova.lmerModLmerTest.

Value

an object containing design name, design data frame, model formula, and a pseudo model object with the expected effect size and variance-covariance components of random effects.