Create an Randomized Complete Block Design

Create an Randomized Complete Block Design

Usage

designRCBD(treatments, label, blocks, formula, beta, VarCov, sigma2, ...)

Arguments

treatments

a vector specifying the treatment structure, in which the length of the vector indicates the number of treatment factors, and each value represents the number of levels for each factor. For example, treatments=c(2,3) specifies two treatment factors with 2 and 3 levels, respectively.

label

Optional. A list specifying the names of treatment factors and their corresponding level labels. Each element of the list represents a factor, with the name of the element being the name of the factor. The value of each element should be a vector containing the labels for that factor's levels. By default, the name of one treatment factor is "trt," and the names of two treatment factors are "facA" and "facB," respectively. The levels of each factor are coded as "1", "2", ..., if not specified in label. For example, list(trt = c("A1", "A2")) assigns the name "trt" to the treatment factor, with its levels labeled as "A1" and "A2." For multiple factors, such as list(A = c("A1", "A2"), B = c("B1", "B2", "B3")), the first factor is named "A" with levels "A1" and "A2," and the second factor is named "B" with levels "B1," "B2," and "B3."

blocks

the number of blocks

formula

a symbolic description of the model that would be used to test the treatment effects in post-experimental data analysis. The way of writing formula , see lm for fixed effects and lmer for random effects. The default formula for one treatment factor and two treatment factors are y ~ trt + (1|block) and y ~ facA*facB +(1|block), respectively, and when the label argument is supplied, the names of treatment factors are replaced by the corresponding names specified in label. The response variable in the formula is always denoted as y, and block factor is always denoted as block.

beta

a vector of the expectations of model coefficients.The coefficients for categorical variables are the coefficients of dummy variables created using contr.treatment contrast coding.

VarCov

variance-covariance components of random effects. If there are multiple random effects of one grouping factor, provide the variance-covariance matrix. If there are multiple grouping factors, supply the variance-covariance matrix of each grouping factor in a list.

sigma2

error variance

...

other arguments passed to the anova function in lmerTest. The type of ANOVA table, with Type III as the default, and the method for computing the denominator degrees of freedom, with Satterthwaite's method as the default, can be changed. For more details, see anova.lmerModLmerTest.

Value

an object containing design name, design data frame, model formula, and a pseudo model object with the expected effect size and variance-covariance components of random effects.