Create data frame for Latin square design

Create data frame for Latin square design

Usage

df.lsd(treatments, squares = 1, reuse = c("row", "col", "both"))

Arguments

treatments

a vector representing the number of levels of treatment factors sequentially. For example, treatments=c(2,3) specifies two treatment factors with 2 and 3 levels, respectively.

squares

number of replicated squares

reuse

a character string: "row", "col", or "both", indicating reuse of rows or columns or both when replicate a Latin square

Value

a data.frame with columns for treatment factors, row and column block factors, and squares

Examples

df.lsd(5, 3)
#>    row col treatment square
#> 1    1   1         2      1
#> 2    2   1         3      1
#> 3    3   1         4      1
#> 4    4   1         5      1
#> 5    5   1         1      1
#> 6    1   2         3      1
#> 7    2   2         4      1
#> 8    3   2         5      1
#> 9    4   2         1      1
#> 10   5   2         2      1
#> 11   1   3         4      1
#> 12   2   3         5      1
#> 13   3   3         1      1
#> 14   4   3         2      1
#> 15   5   3         3      1
#> 16   1   4         5      1
#> 17   2   4         1      1
#> 18   3   4         2      1
#> 19   4   4         3      1
#> 20   5   4         4      1
#> 21   1   5         1      1
#> 22   2   5         2      1
#> 23   3   5         3      1
#> 24   4   5         4      1
#> 25   5   5         5      1
#> 26   1   6         2      2
#> 27   2   6         3      2
#> 28   3   6         4      2
#> 29   4   6         5      2
#> 30   5   6         1      2
#> 31   1   7         3      2
#> 32   2   7         4      2
#> 33   3   7         5      2
#> 34   4   7         1      2
#> 35   5   7         2      2
#> 36   1   8         4      2
#> 37   2   8         5      2
#> 38   3   8         1      2
#> 39   4   8         2      2
#> 40   5   8         3      2
#> 41   1   9         5      2
#> 42   2   9         1      2
#> 43   3   9         2      2
#> 44   4   9         3      2
#> 45   5   9         4      2
#> 46   1  10         1      2
#> 47   2  10         2      2
#> 48   3  10         3      2
#> 49   4  10         4      2
#> 50   5  10         5      2
#> 51   1  11         2      3
#> 52   2  11         3      3
#> 53   3  11         4      3
#> 54   4  11         5      3
#> 55   5  11         1      3
#> 56   1  12         3      3
#> 57   2  12         4      3
#> 58   3  12         5      3
#> 59   4  12         1      3
#> 60   5  12         2      3
#> 61   1  13         4      3
#> 62   2  13         5      3
#> 63   3  13         1      3
#> 64   4  13         2      3
#> 65   5  13         3      3
#> 66   1  14         5      3
#> 67   2  14         1      3
#> 68   3  14         2      3
#> 69   4  14         3      3
#> 70   5  14         4      3
#> 71   1  15         1      3
#> 72   2  15         2      3
#> 73   3  15         3      3
#> 74   4  15         4      3
#> 75   5  15         5      3