Create data frame for split-plot design

Create data frame for split-plot design

Usage

df.splitplot(trt.main, trt.sub, replications)

Arguments

trt.main

a vector representing the number of levels of main plot treatment factors sequentially.

trt.sub

a vector representing the number of levels of sub plot treatment factors sequentially.

replications

number of replications at main plots

Value

a data.frame with columns for main plots, main treatments, and sub-treatments

Examples

df.splitplot(3, 4, 15)
#>     mainplots trt.main trt.sub
#> 1           1        1       1
#> 2           1        1       2
#> 3           1        1       3
#> 4           1        1       4
#> 5           2        2       1
#> 6           2        2       2
#> 7           2        2       3
#> 8           2        2       4
#> 9           3        3       1
#> 10          3        3       2
#> 11          3        3       3
#> 12          3        3       4
#> 13          4        1       1
#> 14          4        1       2
#> 15          4        1       3
#> 16          4        1       4
#> 17          5        2       1
#> 18          5        2       2
#> 19          5        2       3
#> 20          5        2       4
#> 21          6        3       1
#> 22          6        3       2
#> 23          6        3       3
#> 24          6        3       4
#> 25          7        1       1
#> 26          7        1       2
#> 27          7        1       3
#> 28          7        1       4
#> 29          8        2       1
#> 30          8        2       2
#> 31          8        2       3
#> 32          8        2       4
#> 33          9        3       1
#> 34          9        3       2
#> 35          9        3       3
#> 36          9        3       4
#> 37         10        1       1
#> 38         10        1       2
#> 39         10        1       3
#> 40         10        1       4
#> 41         11        2       1
#> 42         11        2       2
#> 43         11        2       3
#> 44         11        2       4
#> 45         12        3       1
#> 46         12        3       2
#> 47         12        3       3
#> 48         12        3       4
#> 49         13        1       1
#> 50         13        1       2
#> 51         13        1       3
#> 52         13        1       4
#> 53         14        2       1
#> 54         14        2       2
#> 55         14        2       3
#> 56         14        2       4
#> 57         15        3       1
#> 58         15        3       2
#> 59         15        3       3
#> 60         15        3       4
#> 61         16        1       1
#> 62         16        1       2
#> 63         16        1       3
#> 64         16        1       4
#> 65         17        2       1
#> 66         17        2       2
#> 67         17        2       3
#> 68         17        2       4
#> 69         18        3       1
#> 70         18        3       2
#> 71         18        3       3
#> 72         18        3       4
#> 73         19        1       1
#> 74         19        1       2
#> 75         19        1       3
#> 76         19        1       4
#> 77         20        2       1
#> 78         20        2       2
#> 79         20        2       3
#> 80         20        2       4
#> 81         21        3       1
#> 82         21        3       2
#> 83         21        3       3
#> 84         21        3       4
#> 85         22        1       1
#> 86         22        1       2
#> 87         22        1       3
#> 88         22        1       4
#> 89         23        2       1
#> 90         23        2       2
#> 91         23        2       3
#> 92         23        2       4
#> 93         24        3       1
#> 94         24        3       2
#> 95         24        3       3
#> 96         24        3       4
#> 97         25        1       1
#> 98         25        1       2
#> 99         25        1       3
#> 100        25        1       4
#> 101        26        2       1
#> 102        26        2       2
#> 103        26        2       3
#> 104        26        2       4
#> 105        27        3       1
#> 106        27        3       2
#> 107        27        3       3
#> 108        27        3       4
#> 109        28        1       1
#> 110        28        1       2
#> 111        28        1       3
#> 112        28        1       4
#> 113        29        2       1
#> 114        29        2       2
#> 115        29        2       3
#> 116        29        2       4
#> 117        30        3       1
#> 118        30        3       2
#> 119        30        3       3
#> 120        30        3       4
#> 121        31        1       1
#> 122        31        1       2
#> 123        31        1       3
#> 124        31        1       4
#> 125        32        2       1
#> 126        32        2       2
#> 127        32        2       3
#> 128        32        2       4
#> 129        33        3       1
#> 130        33        3       2
#> 131        33        3       3
#> 132        33        3       4
#> 133        34        1       1
#> 134        34        1       2
#> 135        34        1       3
#> 136        34        1       4
#> 137        35        2       1
#> 138        35        2       2
#> 139        35        2       3
#> 140        35        2       4
#> 141        36        3       1
#> 142        36        3       2
#> 143        36        3       3
#> 144        36        3       4
#> 145        37        1       1
#> 146        37        1       2
#> 147        37        1       3
#> 148        37        1       4
#> 149        38        2       1
#> 150        38        2       2
#> 151        38        2       3
#> 152        38        2       4
#> 153        39        3       1
#> 154        39        3       2
#> 155        39        3       3
#> 156        39        3       4
#> 157        40        1       1
#> 158        40        1       2
#> 159        40        1       3
#> 160        40        1       4
#> 161        41        2       1
#> 162        41        2       2
#> 163        41        2       3
#> 164        41        2       4
#> 165        42        3       1
#> 166        42        3       2
#> 167        42        3       3
#> 168        42        3       4
#> 169        43        1       1
#> 170        43        1       2
#> 171        43        1       3
#> 172        43        1       4
#> 173        44        2       1
#> 174        44        2       2
#> 175        44        2       3
#> 176        44        2       4
#> 177        45        3       1
#> 178        45        3       2
#> 179        45        3       3
#> 180        45        3       4