Create an artificial model object

Create a pseudo-model object with the response variable being simulated according to the fixed and random effects. Model coefficients are replaced by the expectations specified in the argument beta. Variance-covariance components of random effects are replaced by the values specified in argument VarCov. The standard deviation of random error is replaced by the argument sigma. Creating such a pseudo-model facilitates power calculations by leveraging the anova function in lmerTest and the Anova function in car.

Usage

fit.pseu.model(formula, data, beta, VarCov, sigma, ...)

Arguments

formula

an object of class formula: a symbolic description of the model that would be used to test effects in post-experimental data analysis. For details on model specification, see lm for fixed effects and lmer for random effects.

data

a data frame with the independent variables of a design as columns, e.g., treatment factors and block factors.

beta

a vector of the expectations of model coefficients. The coefficients for categorical variables are the coefficients of dummy variables created using contr.treatment contrast coding.

VarCov

variance-covariance components of random effects. If there are multiple random effects of one grouping factor, provide the variance-covariance matrix. If there are multiple grouping factors, supply the variance-covariance matrix of each grouping factor in a list.

sigma

standard deviation of error

...

other arguments passed to the anova function in lmerTest. The type of ANOVA table, with Type III as the default, and the method for computing the denominator degrees of freedom, with Satterthwaite's method as the default, can be changed. For more details, see anova.lmerModLmerTest.

Value

a pseudo model object.